Move Fields to Stdlib/Hydrator
